fix: optimization and table bug fix

+++ core.py
@@ -1,3 +1,5 @@
+from os.path import isfile
+
 from sqlalchemy import create_engine
 
 from models import (
@@ -15,14 +17,12 @@ from models import (
 
 
 def create_sql_tables():
-    engine = create_engine("sqlite:///DB/DataBase.sqlite", echo=False)
-    metadata_obj.drop_all(engine)
-    metadata_obj.create_all(engine)
-    create_table(Nir_NTP, NTP)
-    create_table(Nir_Grant, Grant)
-    create_table(Nir_Templan, Templan)
-    create_table(Nir_VUZ, VUZ)
+    if not(isfile('DB/DataBase.sqlite')):
+        engine = create_engine("sqlite:///DB/DataBase.sqlite", echo=False)
+        metadata_obj.drop_all(engine)
+        metadata_obj.create_all(engine)
+        create_table(Nir_NTP, NTP)
+        create_table(Nir_Grant, Grant)
+        create_table(Nir_Templan, Templan)
+        create_table(Nir_VUZ, VUZ)
 
-
-if __name__ == "__main__":
-    create_sql_tables()
